scripts/requirements.txt
mysql-connector-python==8.0.29
mysql-connector-repackaged==0.3.1
python-dotenv==0.20.0

